Don't start vncserver if its already running
Problem: deploy failure, vncserver can not be started if its already running
on a display
Solution: Use a command to test if vnc is already running. Then start it if
it's not running.

To elaborate on the command used..
vncserver (running in the guest vm) will listen for connections on port 5900 + $DISPLAY according to
https://guacamole.incubator.apache.org/doc/gug/configuring-guacamole.html#vnc-network-parameters

Since we use the default port and display 5, it should be reliable.