Fix instances fail to boot if a volume is removed outside atmosphere
Problem:
If a volume is removed outside of atmosphere, the instance will fail to boot
after being rebooted/redeployed/resumed

Solution:
Use options from fstab(5) to prevent missing disks from failing the boot
process

This PR adds the 'nobootwait' or 'nofail' option to the `/etc/fstab` line of
the mounted disk. This allows instances to boot even if the volume is missing.
'nofail' is used on all instances except Ubuntu 14 instances which use
'nobootwait'. These fstab changes are made for ephemeral disks and mounted
volumes.