Link instance-scripts so they are inspectable
Increase the debuggability of instance-scripts by allowing them to
remain available for inspection after the instance deployment activity
is complete.

This work makes the deploy scripts available under:
- /etc/atmo/instance-scripts-avaiable/

---

Terminology:
------------

We will refer to scripts associated with an instance at the time of
"launch" as "deploy scripts", not *boot scripts*.

For Deploy Scripts, this "assocation" at launch time is the most common.
A deploy script *could* be _included_ with an Image by the author of
that Image. For this commit (and the containing PR), we are not trying
to handle nor account for deploy scripts included in Images. This case
is referred to by Steve Gregory (the case of deploy scripts included
"with an Image") as "Post Scripts". These deploy scripts included with
an Image will always be run **last** - which seems to be why they were
called this. We will endeavor for a "new" name for them at some point.
At this time - they do not appear to be working (as expected) [0].

---

Description:

This links scripts associated at instance "launch time" into a directory
that is _inspectable_ after execution. It borrows from the pattern that
you see in uWSGI & Nginx regarding "apps" and "sites (respectively).

For Nginx, you have a site's configuration "sites-available" and when it
is active (or enabled), the site's configuration is symlinked into the
"sites-enabled" directory:

```
ls -lha /etc/nginx/sites-available/
total 16K
drwxr-xr-x 2 root root 4.0K Jan 19 21:53 .
drwxr-xr-x 9 root root 4.0K Jan 19 21:53 ..
-rw-r--r-- 1 root root 2.4K Jan 18  2017 default
-rw-r--r-- 1 root root 1.9K Jan 19 19:11 site.conf
```

```
ls -lha /etc/nginx/sites-enabled/
total 8.0K
drwxr-xr-x 2 root root 4.0K Nov 13 23:34 .
drwxr-xr-x 9 root root 4.0K Jan 19 21:53 ..
lrwxrwxrwx 1 root root   36 Nov 13 23:34 site.conf -> /etc/nginx/sites-available/site.conf
```

Here - we take the "deploy scripts" copy them into the "available"
directory:

- /etc/atmo/instance-scripts-avaiable/

Them, we symlink them into location the Ansible Role expects them to be:

- /etc/atmo/instance-scripts.d/

This approach of removing from "instance-scripts.d" was used to help
avoid any of "re-executing" scripts.

[0] https://pods.iplantcollaborative.org/jira/browse/ATMO-2137