Fix instances that were failing to get networking
Problem:
Restarting networking during a system's boot would interfere with the system
getting networking.

Solution:
Don't perform a service network restart

We were performing a restart because it seemed like a reliable to ensure a new
dhcp lease. We're reverting to our original behavior of running dhclient, but
this time we don't fail if dhclient is already running/getting a lease