Fix incorrect condition
Problem:
Ansible would fail on ubuntu 14 with: Could not find the requested service sshd: host

Solution:
The error wasn't super clear, but the issue was that we were retrying to
restart `sshd` on ubuntu systems which is wrong. It should be retrying to
restart `ssh`. It turned out CentOS variables we're being loaded
everytime even for ubuntu systems. This is because the when condition was
always true.

In the below task: item.condition expands to the string: "ansible_distribution == ...".
In the when clause we're then asking if that string is truthy, which it always is.

    - name: Gather OS-specific variables
      include_vars: "{{ item.var }}"
      when: "item.condition"
      with_items:
        - { var: 'Ubuntu.yml', condition: ansible_distribution == "Ubuntu" }
        - { var: 'CentOS.yml', condition: ansible_distribution == "CentOS" }

You might expect item.condition to have been true or false, but in yaml
everything is a string unless your in a context where values are evaluated
like between double brackets or in a when.